Move prepareOptions to test helpers
