+++ test/helpers/prepareOptions.js
